[MODIFY-FRONTEND] RootGuard added to _layout.jsx + missed payments.jsx
