+++ frontend/app/(tabs)/payments.jsx
@@ -1,0 +1,9 @@
+import { Text, View } from 'react-native';
+
+export default function Payments() {
+  return (
+    <View style={{ flex: 1, justifyContent: 'center', alignItems: 'center' }}>
+      <Text>Payments</Text>
+    </View>
+  );
+}
